feat: add some api

+++ web-voyager-ddd-const/src/main/java/io/github/sztuwebclub/webvoyager/constant/utils/ContextUtil.java
@@ -3,13 +3,13 @@ package io.github.sztuwebclub.webvoyager.constant.utils;
 public class ContextUtil {
     private ContextUtil() {}
 
-    public static final ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static final ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
